Test build with esm, node projects.
